requirements.txt
alembic==1.14.0
blinker==1.9.0
click==8.1.7
dnspython==2.7.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pymongo==4.10.1
PyMySQL==1.1.1
pytest==8.3.3
pytest-flask==1.3.0
python-dotenv==1.0.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
